This commit fixes issue 1.
When editing a previous choice entry, Angularjs was unable to update
the model as Javascript strings are immutable. The solution was to
use an array of objects for the choices model with a single key-value
pair (text:string). Upon creating the poll, unwrap this array of objects
before sending the JSON request to the server.